+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Yii Framework 2 bootstrap4 extension Change Log
 ==============================================
 
+2.0.4 under development
+-----------------------
+
+- no changes in this release.
+
+
 2.0.3 June 04, 2019
 -------------------